update make linux maker, but window does not yet works

+++ make_linux_binary.sh
@@ -1,5 +1,15 @@
-# See https://myridia.com/dev_posts/view/3489 for more info 
-docker run -v "$(pwd):/src/" engineervix/pyinstaller-linux "pyinstaller -F main.py"
-docker run -v "$(pwd):/src/" engineervix/pyinstaller-linux
-sudo cp config.ini dist/
-sudo cp config.ini dist/linux
+#!/bin/sh
+# See https://myridia.com/dev_posts/view/3489 for more info
+
+docker run -v "$(pwd):/src/" batonogov/pyinstaller-linux "pyinstaller -F main.py"
+docker run -v "$(pwd):/src/" batonogov/pyinstaller-linux
+
+if [ -f config.ini ] #check if file exists
+then
+   echo "...copy config.ini to dist/config.ini"    
+   sudo cp config.ini dist/
+else
+   echo "...copy config_defaut.ini to dist/config.ini"
+   sudo cp config_default.ini dist/config.ini
+fi
+
